+++ packages/login_client/lib/src/models/api_consts.dart
@@ -1,0 +1,7 @@
+class ApiConsts {
+  Uri authEndpoint;
+  String authClientId;
+  String authSecret;
+  List<String> authScopes;
+  Uri apiEndpoint;
+}