+++ packages/login_client/lib/src/models/token.dart
@@ -1,0 +1,5 @@
+class Token {
+  String accessToken;
+  String refreshToken;
+  DateTime expiration;
+}