+++ packages/login_client/lib/src/strategies/strategies.dart
@@ -1,0 +1,4 @@
+export './assertion_strategy.dart';
+export './password_credentials_strategy.dart';
+export './preauthentication_strategy.dart';
+export './sms_token_strategy.dart';